Move integration fixtures
it looks like out-of-process stuff is having trouble finding them. The CI should be adding `tests/` to the universal python path though, so my hope is `static` is then findable
